+++ src/ScriptEngine/Machine/IValue.cs
@@ -10,7 +10,7 @@ namespace ScriptEngine.Machine
         DataType DataType { get; }
         TypeDescriptor SystemType { get; }
         
-        double AsNumber();
+        decimal AsNumber();
         DateTime AsDate();
         bool AsBoolean();
         string AsString();
